backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiomysql
aiosqlite
alembic
celery[redis]
redis
boto3
pytesseract
Pillow
httpx
sendgrid
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
python-multipart
jinja2
flower
pytest
pytest-asyncio
aiofiles
email-validator
